+++ thoth/storages/graph/__init__.py
@@ -18,5 +18,4 @@
 """A graph database adapter for communicating with dgraph via gRPC."""
 
 
-from .cache import GraphCache
 from .postgres import GraphDatabase